Centralize test constants, always release on logout, normalize ARNs at source
